Specialize accounts and related data to EVMNetwork
On our journey to full multi-network support, we need to be able to rely
on chain IDs throughout much of the wallet. Until we have a
comprehensive way to refer to different chains consistently, narrowing
AddressNetwork saves quite a bit of type accounting.

This applies to AddressNetwork, NameNetwork, and AccountBalance.

Co-authored-by: Matt Luongo <mhluongo@gmail.com>